Cleaning up some formatting
and moving a method from one messaging.Util to app.Util because it specifies default profile which is in the display and UI domain
