+++ app/src/main/java/io/islnd/android/islnd/app/util/Util.java
@@ -200,4 +200,14 @@ public class Util
 
         qrImageView.setImageBitmap(bmp);
     }
+
+    public static Profile buildDefaultProfile(Context context, String username) {
+        // TODO: default image Uris will probably be assets...
+        return new Profile(
+                username,
+                context.getString(R.string.profile_default_about_me),
+                ImageUtil.getDefaultProfileImageUri(context),
+                ImageUtil.getDefaultHeaderImageUri(context),
+                Integer.MIN_VALUE);
+    }
 }
