+++ app/src/main/java/io/islnd/android/islnd/messaging/Util.java
@@ -29,14 +29,4 @@ public class Util {
 
         return newestItem;
     }
-
-    public static Profile buildDefaultProfile(Context context, String username) {
-        // TODO: default image Uris will probably be assets...
-        return new Profile(
-                username,
-                context.getString(R.string.profile_default_about_me),
-                ImageUtil.getDefaultProfileImageUri(context),
-                ImageUtil.getDefaultHeaderImageUri(context),
-                Integer.MIN_VALUE);
-    }
 }
